user_service/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==5.0.1
hiredis==2.2.3

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyotp==2.9.0
qrcode[pil]==7.4.2

# OAuth
authlib==1.2.1
httpx==0.25.1

# Password validation
zxcvbn==4.4.28

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Cryptography and KMS
cryptography==41.0.7
boto3==1.29.7  # AWS KMS
hvac==2.0.0   # HashiCorp Vault

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pycountry==23.12.11

# Rate limiting
slowapi==0.1.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1
faker==20.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# Background tasks
apscheduler==3.10.4
